Fix issues around navigation
