+++ web-common/src/features/models/navigation/ModelMenuItems.svelte
@@ -1,11 +1,6 @@
 <script lang="ts">
   import Explore from "@rilldata/web-common/components/icons/Explore.svelte";
-  import {
-    getFileAPIPathFromNameAndType,
-    getFilePathFromNameAndType,
-  } from "@rilldata/web-common/features/entity-management/entity-mappers";
   import { fileArtifacts } from "@rilldata/web-common/features/entity-management/file-artifacts";
-  import { EntityType } from "@rilldata/web-common/features/entity-management/types";
   import { featureFlags } from "@rilldata/web-common/features/feature-flags";
   import NavigationMenuItem from "@rilldata/web-common/layout/navigation/NavigationMenuItem.svelte";
   import { BehaviourEventMedium } from "@rilldata/web-common/metrics/service/BehaviourEventTypes";
@@ -15,52 +10,38 @@
   import { createEventDispatcher } from "svelte";
   import { V1ReconcileStatus } from "../../../runtime-client";
   import { runtime } from "../../../runtime-client/runtime-store";
-  import { deleteFileArtifact } from "../../entity-management/actions";
   import { useCreateDashboardFromTableUIAction } from "../../metrics-views/ai-generation/generateMetricsView";
-  import { useModel, useModelRoutes } from "../selectors";
 
-  export let modelName: string;
+  export let filePath: string;
 
-  $: modelPath = getFilePathFromNameAndType(modelName, EntityType.Model);
-  $: fileArtifact = fileArtifacts.getFileArtifact(modelPath);
+  $: fileArtifact = fileArtifacts.getFileArtifact(filePath);
 
   const queryClient = useQueryClient();
   const dispatch = createEventDispatcher();
 
   const { customDashboards } = featureFlags;
 
-  $: modelRoutes = useModelRoutes($runtime.instanceId);
   $: modelHasError = fileArtifact.getHasErrors(
     queryClient,
     $runtime.instanceId,
   );
-  $: modelQuery = useModel($runtime.instanceId, modelName);
+  $: modelQuery = fileArtifact.getResource(queryClient, $runtime.instanceId);
   $: connector = $modelQuery.data?.model?.spec?.connector;
   $: modelIsIdle =
     $modelQuery.data?.meta?.reconcileStatus ===
     V1ReconcileStatus.RECONCILE_STATUS_IDLE;
   $: disableCreateDashboard = $modelHasError || !modelIsIdle;
+  $: tableName = $modelQuery.data?.model?.state?.table ?? "";
 
   $: createDashboardFromTable = useCreateDashboardFromTableUIAction(
     $runtime.instanceId,
     connector as string,
     "",
     "",
-    modelName,
+    tableName,
     BehaviourEventMedium.Menu,
     MetricsEventSpace.LeftPanel,
   );
-
-  const handleDeleteModel = async (modelName: string) => {
-    if ($modelRoutes.data) {
-      await deleteFileArtifact(
-        $runtime.instanceId,
-        getFileAPIPathFromNameAndType(modelName, EntityType.Model),
-        EntityType.Model,
-        $modelRoutes.data,
-      );
-    }
-  };
 </script>
 
 <NavigationMenuItem
